Revert "Merge NYT to master"
