+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.androidfinalprojectw18">
 
+    <uses-permission android:name="android.permission.INTERNET" ></uses-permission>
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -9,10 +11,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ArticleSearchNYT"></activity>
+        <activity android:name=".websterdictionary.ViewDictionaryItem"></activity>
+        <activity android:name=".ArticleSearchNYT" />
         <activity android:name=".FlightStatusTracker" />
         <activity android:name=".NewsFeed" />
-        <activity android:name=".MerriamWebsterDictionary" />
+        <activity android:name=".websterdictionary.MerriamWebsterDictionary" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
