--- app/src/main/java/com/example/androidfinalprojectw18/ArticleModel.java
@@ -1,64 +1,0 @@
-package com.example.androidfinalprojectw18;
-
-public class ArticleModel {
-
-    public String title;
-    public String author;
-    public String description;
-    public String date;
-    public String link;
-    public String imgURL;
-
-
-
-    public ArticleModel() {
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getDate() {
-        return date;
-    }
-
-    public void setDate(String date) {
-        this.date = date;
-    }
-
-    public String getLink() {
-        return link;
-    }
-
-    public void setLink(String link) {
-        this.link = link;
-    }
-
-    public String getImgURL() {
-        return imgURL;
-    }
-
-    public void setImgURL(String imgURL) {
-        this.imgURL = imgURL;
-    }
-}
