+++ app/src/main/res/layout/activity_article_search_nyt.xml
@@ -1,27 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent" android:layout_height="match_parent">
-
-    <include
-        android:id="@+id/nytimes_toolbar"
-        layout="@layout/nytimes_toolbar" />
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ArticleSearchNYT"
+    android:orientation="vertical">
 
     <ListView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/listView"
-        android:layout_below="@id/nytimes_toolbar"
-        >
+        android:layout_height="wrap_content">
+
     </ListView>
 
-    <ProgressBar
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="invisible"
-        android:id="@+id/progressBar"
-        style="?android:attr/progressBarStyleHorizontal"/>
-
-
-</RelativeLayout>
+</LinearLayout>
