--- app/src/main/res/layout/home_activity.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:orientation="vertical"
-    android:weightSum="80"
-    android:background="#006341">
-
-
-    <ImageView
-        android:id="@+id/imageView2"
-        android:layout_marginTop="50dp"
-        android:layout_width="300dp"
-        android:layout_height="200dp"
-        android:layout_centerHorizontal="true"
-        app:srcCompat="@drawable/ac_logo1" />
-
-    <Button
-        android:id="@+id/flight_btn"
-        android:layout_marginTop="20dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:background="@drawable/flight_tracker_btn"
-        android:layout_below="@id/imageView2"
-        android:onClick="launchActivity"/>
-
-    <Button
-        android:id="@+id/feed_btn"
-        android:layout_marginTop="20dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:background="@drawable/newsfeed_btn"
-        android:layout_below="@id/flight_btn"
-        android:onClick="launchActivity"/>
-    <Button
-        android:id="@+id/dictionary_btn"
-        android:layout_marginTop="20dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:background="@drawable/dictionary_btn"
-        android:layout_below="@id/feed_btn"
-        android:onClick="launchActivity"/>
-
-    <Button
-        android:id="@+id/search_btn"
-        android:layout_marginTop="20dp"
-        android:layout_marginLeft="20dp"
-        android:layout_marginRight="20dp"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:background="@drawable/nytimes_btn"
-        android:layout_below="@id/dictionary_btn"
-        android:onClick="launchActivity"/>
-
-</RelativeLayout>
