--- app/src/main/res/layout/search_dialog.xml
@@ -1,14 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-
-    <EditText android:id="@+id/searchInput"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
-
-
-
-
-</LinearLayout>
