--- app/src/main/res/menu/nytimes_menu.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-
-
-    <item android:id="@+id/MenuItems-search"
-        android:icon="@drawable/search"
-        android:orderInCategory="1"
-        app:showAsAction="always"/>
-    <item android:id="@+id/MenuItems-bookmark"
-        android:icon="@drawable/bookmark"
-        android:orderInCategory="2"
-        app:showAsAction="always"/>
-
-</menu>
